Refactored code. Split algorithms in Converter into different files. Added docs just everywhere. Few improvements.
