+++ project/Converter_Huffman.cpp
@@ -1,0 +1,25 @@
+#include "Converter.h"
+
+
+/** \brief Donverting method for Huffman mode
+ *
+ * \param filename const std::string& converted file path
+ * \return void
+ *
+ */
+void Converter::conv_huffman(const std::string& filename)
+{
+    // TODO
+}
+
+
+/** \brief Deconverting method for Huffman mode
+ *
+ * \param filename const std::string& converted file path
+ * \return void
+ *
+ */
+void Converter::dconv_huffman(const std::string& filename)
+{
+    // TODO
+}
