+++ project/Converter_RLE.cpp
@@ -1,0 +1,25 @@
+#include "Converter.h"
+
+
+/** \brief Converting method for RLE mode
+ *
+ * \param filename const std::string& converted file path
+ * \return void
+ *
+ */
+void Converter::conv_rle(const std::string& filename)
+{
+    // TODO
+}
+
+
+/** \brief Deconverting method for RLE mode
+ *
+ * \param filename const std::string& converted file path
+ * \return void
+ *
+ */
+void Converter::dconv_rle(const std::string& filename)
+{
+    // TODO
+}
