+++ project/SDL_function.h
@@ -7,11 +7,11 @@
 
 namespace SDL
 {
-    Uint32 get_pixel(SDL_Surface *surface, int x, int y);   // get pixel from SDL_surface
-    uint8_t* get_pixel2(SDL_Surface* surf, int x, int y);
-    void draw_pixel(SDL_Surface *surface, int x, int y, uint8_t R, uint8_t G, uint8_t B);
+    Uint32 get_pixel(SDL_Surface *surface, const int& x, const int& y);   // get pixel from SDL_surface
+    uint8_t* get_pixel2(SDL_Surface* surf, const int& x, const int& y);
+    void draw_pixel(SDL_Surface *surface, const int& x, const int& y, const uint8_t& R, const uint8_t& G, const uint8_t& B);
     SDL_Surface* new_bmp_surface(const std::string& filename);
-    SDL_Surface* new_empty_surface(int width, int height);
+    SDL_Surface* new_empty_surface(const int& width, const int& height);
     void delete_surface(SDL_Surface* surf);
 }
