+++ project/Window.h
@@ -10,11 +10,11 @@
 
 class Window
 {
-    friend class Converter;
     public:
         using uint = unsigned int;
 
         explicit Window(std::string filename);
+        explicit Window(SDL_Surface* image);
         ~Window();
         Window(const Window&) = delete;
         Window(Window&&) = delete;
@@ -25,21 +25,18 @@ class Window
         uint m_height;
         std::string m_filename;
 
-        void set_dimensions();  // m_width and m_height init
-
-    // SDL STUFF ////////////
         SDL_Surface* m_screen;
         SDL_Surface* m_bmp;
         SDL_Rect* m_dstrect; // for centering bitmap on the screen
         void init();
         void create_window();
         void load_image();
+        void set_dimensions();  // m_width and m_height init
     public:
         void draw_image();
 
         inline uint get_width() { return m_width; };
         inline uint get_height() { return m_height; };
-        void get_pixel(int x, int y, int& R, int& G, int& B);
 };
 
 #endif // WINDOW_H___
